requirements.txt
requests==2.31.0
python-dotenv==1.0.1
plotly==5.20.0
pandas==2.2.2
matplotlib==3.9.0
flask==3.0.3